+++ src/test/kotlin/io/kjson/ktor/JSONKtorReturnTest.kt
@@ -1,7 +1,7 @@
 /*
  * @(#) JSONKtorReturnTest.kt
  *
- * kjson-ktor  Reflection-based JSON serialization and deserialization for ktor
+ * kjson-ktor  Reflection-based JSON serialization and deserialization for Ktor
  * Copyright (c) 2023 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy