+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [0.3] - 2023-05-21
+### Added
+- `CoOutputChannel`
+- `JSONKtorClientException`
+### Changed
+- `JSONKtorFunctions`: added `channelOutput()`, `createStreamedJSONContent()` and `ApplicationCall.respondStream()`
+- `README.md`: more documentation
+
+## [0.2] - 2023-05-14
+### Added
+- `KtorByteChannelOutput`
+- `KtorOutgoingContent`
+### Changed
+- `JSONKtorFunctions`: added `channelOutput()` and `ApplicationCall.respondStream()`
+
 ## [0.1] - 2023-04-25
 ### Added
 - all files: initial versions